+++ lib/linters.rb
@@ -1,4 +1,11 @@
-class Linters 
+class Linters
+    attr_reader :file_path, :content, :file_lines, :errors
+    def initialize(file_path)
+        @file_path = file_path
+        @file = file_path.to_s
+        @file_lines = File.readlines(@file)
+        @errors = []
+    end
 
     def trailing_white_spaces
         
@@ -26,9 +33,6 @@ class Linters
             @errors << 'You are missing "end"'
         end
     end
-
-
-
 end
